+++ src/postgres/commands/enterPostgresCredentials.ts
@@ -8,6 +8,7 @@ import { IActionContext } from "vscode-azureextensionui";
 import { ext } from "../../extensionVariables";
 import { localize } from "../../utils/localize";
 import { nonNullProp } from '../../utils/nonNull';
+import { PostgresServerType } from '../abstract/models';
 import { PostgresServerTreeItem } from "../tree/PostgresServerTreeItem";
 import { setPostgresCredentials } from "./setPostgresCredentials";
 
@@ -18,17 +19,21 @@ export async function enterPostgresCredentials(context: IActionContext, treeItem
 
     let username: string = await context.ui.showInputBox({
         prompt: localize('enterUsername', 'Enter username for server "{0}"', treeItem.label),
+        stepName: 'enterPostgresUsername',
         validateInput: (value: string) => { return (value && value.length) ? undefined : localize('usernameCannotBeEmpty', 'Username cannot be empty.'); }
     });
 
     const serverName: string = nonNullProp(treeItem, 'azureName');
+    // Username doesn't contain servername prefix for Postgres Flexible Servers only
+    // As present on the portal for any Flexbile Server instance
     const usernameSuffix: string = `@${serverName}`;
-    if (!username.includes(usernameSuffix)) {
+    if (treeItem.serverType === PostgresServerType.Single && !username.includes(usernameSuffix)) {
         username += usernameSuffix;
     }
 
     const password: string = await context.ui.showInputBox({
         prompt: localize('enterPassword', 'Enter password for server "{0}"', treeItem.label),
+        stepName: 'enterPostgresPassword',
         password: true,
         validateInput: (value: string) => { return (value && value.length) ? undefined : localize('passwordCannotBeEmpty', 'Password cannot be empty.'); }
     });